Crawler/requirements.txt
firecrawl-py>=1.0.0
python-dotenv>=1.0.0

